on second thought, move downloader
